requirements.txt
asgiref==3.5.1
dj-database-url==0.5.0
Django==4.0.4
django-heroku==0.0.0
djangorestframework==3.13.1
gunicorn==20.1.0
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.1.0
